Extract some code into a separate class, and some other code into separate methods
